RC2 Working SaveAs/Revert/SaveCopy Fixed cell still selected bug. Major cleanup

+++ FractalScape/MDBFractalLibraryCollectionSource.m
@@ -52,7 +52,7 @@
             }
         }
         NSString* stackString = [filteredStack debugDescription];
-        NSLog(@"Count: %lu Stack: %@",count, stackString);
+        NSLog(@"Count: %lu Stack: %@",(unsigned long)count, stackString);
     }
     
     return count;
